Better aliases; Better install_packages function; Better usages

+++ lib/install.sh
@@ -8,26 +8,28 @@
 
 packages_to_install=" "
 
-install() {
+required() {
     echo "$packages_to_install" | grep " $1 " > /dev/null && return 0
     return 1
 }
 
 install_packages() {    
     # vscode
-    install "code" && {
-        code --version > /dev/null
-        [ $? -eq 0 ] || {
+    required "code" && {
+        # test if installed
+        if ! command -v "code" &> /dev/null; then
             wget -qO- https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor > ~/packages.microsoft.gpg
             sudo install -o root -g root -m 644 ~/packages.microsoft.gpg /etc/apt/trusted.gpg.d/
             rm ~/packages.microsoft.gpg
             sudo sh -c 'echo "deb [arch=amd64 signed-by=/etc/apt/trusted.gpg.d/packages.microsoft.gpg] https://packages.microsoft.com/repos/vscode stable main" > /etc/apt/sources.list.d/vscode.list'
             sudo apt-get install apt-transport-https
             sudo apt-get update
-        }
+        else
+            echo "code is already installed !"
+        fi
     }
 
-    install "your_package" && {
+    required "your_package" && {
         echo "Do stuff before installing this package"
     }
